payments: reorganize src into lib/ + components/{layout,modals,ui}, extract providers and app-shell
- Move api/wagmi-config/payment-network and utils/* into lib/
- Move layout/ under components/layout/, split modals into components/modals/
- Extract Providers (wagmi + react-query + rainbowkit + authorized-wallet) into providers.tsx
- Extract AppShell from app.tsx into components/layout/app-shell.tsx
- New hooks: use-deposit (approve/deposit state machine), use-tab-url, use-theme
- New SessionExpiredOverlay component
- Move styles/global.scss to src/global.scss
