+++ apps/payments/web/src/components/modals/action-modal.tsx
@@ -1,5 +1,5 @@
 import { useEffect, type ReactNode } from 'react';
-import { useBodyScrollLock } from '../hooks/use-body-scroll-lock';
+import { useBodyScrollLock } from '../../hooks/use-body-scroll-lock';
 
 interface ActionModalProps {
   isOpen: boolean;
