+++ apps/payments/web/src/global.scss
