+++ apps/payments/web/src/hooks/use-theme.ts
@@ -1,0 +1,20 @@
+import { useCallback, useEffect, useState } from 'react';
+
+const STORAGE_KEY = 'antseed-payments-theme';
+
+export function useTheme() {
+  const [isDark, setIsDark] = useState(() => {
+    const saved = localStorage.getItem(STORAGE_KEY);
+    if (saved) return saved === 'dark';
+    return window.matchMedia('(prefers-color-scheme: dark)').matches;
+  });
+
+  useEffect(() => {
+    document.documentElement.setAttribute('data-theme', isDark ? 'dark' : 'light');
+    localStorage.setItem(STORAGE_KEY, isDark ? 'dark' : 'light');
+  }, [isDark]);
+
+  const toggleTheme = useCallback(() => setIsDark((d) => !d), []);
+
+  return { isDark, toggleTheme };
+}
