+++ apps/payments/web/src/lib/api.ts
@@ -1,4 +1,4 @@
-import type { BalanceData, PaymentConfig } from './types';
+import type { BalanceData, PaymentConfig } from '../types';
 
 const BASE = '';
