+++ apps/payments/web/src/lib/payment-network.ts
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { useAccount, useSwitchChain } from 'wagmi';
 import { base } from 'wagmi/chains';
 import { useChainModal } from '@rainbow-me/rainbowkit';
-import type { PaymentConfig } from './types';
+import type { PaymentConfig } from '../types';
 
 const PAYMENT_CHAINS = {
   [base.id]: base,
