+++ apps/payments/web/src/main.tsx
@@ -1,28 +1,12 @@
 import { createRoot } from 'react-dom/client';
-import { WagmiProvider } from 'wagmi';
-import { RainbowKitProvider } from '@rainbow-me/rainbowkit';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { wagmiConfig } from './wagmi-config';
 import { App } from './app';
+import { Providers } from './providers';
 import '@rainbow-me/rainbowkit/styles.css';
-import './styles/global.scss';
-
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      refetchOnWindowFocus: false,
-      retry: 1,
-    },
-  },
-});
+import './global.scss';
 
 const root = document.getElementById('root')!;
 createRoot(root).render(
-  <WagmiProvider config={wagmiConfig}>
-    <QueryClientProvider client={queryClient}>
-      <RainbowKitProvider>
-        <App />
-      </RainbowKitProvider>
-    </QueryClientProvider>
-  </WagmiProvider>
+  <Providers>
+    <App />
+  </Providers>
 );
